Moving /test_utils into the pkg directory
